Potential fix for code scanning alert no. 32

+++ drivers/mmc/arm_pl180_mmci.c
@@ -167,7 +167,7 @@ static int write_bytes(struct mmc *dev, u32 *src, u32 blkcount, u32 blksize)
 {
 	u32 *tempbuff = src;
 	int i;
-	u64 xfercount = blkcount * blksize;
+	u64 xfercount = (u64)blkcount * blksize;
 	struct pl180_mmc_host *host = dev->priv;
 	u32 status, status_err;
